Allow module libs to use VINTF AIDL.
Currently a native APEX can get and use a VINTF AIDL. However, this
can't be passed over JNI to be used by Java code.

This does not open VINTF AIDLs to any type of apps, where AIDL as an API
is completely disallowed. Also, no Java ServiceManager APIs are
available, and they couldn't be until b/139325468 is fixed.

Bug: 161501127
Test: atest binderVendorDoubleLoadTest

Change-Id: I1ab977248b1c39a7d08e0ca34389a7ba168c05b0

Former-commit-id: 95dd53bb5fc8f506e40cc88b1f0eb926bc50c7a5