Screenshot secure layers in ColorFade
When taking a screenshot for ColorFade, allow the screenshot to capture
secure layers, but ensure the SurfaceControl is also marked as secure.
This is to ensure that the screenshot that contains the secure layer
can't itself be screenshotted.

Also cleaned up some unused SurfaceControl.screenshot methods that are
no longer used due to this change

Fixes: 155744498
Test: Turn screen off with secure layers on screen
Change-Id: I064a93f5210af696d52df21380801134b57acc9f

Former-commit-id: 3ce0539bd3422b1cf4fe8cce7b9581e134dc3783