Add onCleanup() method to Pluggables
onCleanup() is called on all pluggables at the end of each pipeline run.
This allows pluggables to clean up any temporary state they may have
created during the run.

We'll use this in a following CL to reset a cache in the
PreparationCoordinator's filter at the end of every run.

Test: atest SystemUITests
Change-Id: I9567aaf50b55aae4241811404856d4225872e972

Former-commit-id: 44c695bb087e1b60f5e3082044f2676a99ddd83b