Remove home snapshot immediately
If the size of snapshot is mismatched, the removal of the starting
window will be deferred. But if the snapshot target is home, since
it is only presented for unlocking, it is better to dismiss it as
soon as possible to avoid outdated content from being shown on screen
if the windows of home have drawn.

Bug: 161530286
Test: Keep home on top and turn off screen with secured lock. Use
      fingerprint to unlock. If home is drawn fast enough, the
      previous state won't show on screen (e.g. the time on clock).
Change-Id: I103f49eb539f78455ad38f48f9cd853dee51b135
(cherry picked from commit 061590c84899ee9ffa9e4188cb20473e5e52a083 [formerly a17c813b4029602fa3f1c2c9baa6455845496106])

Former-commit-id: 106f745c3a13d172c0e4d9935861d060edb4a98f