Change TimeZoneDetectorStrategyImpl to allow geo
Change TimeZoneDetectorStrategyImpl to allow a new "geo only" mode.

The strategy can now exist in one of two modes:
1) Telephony signals only
2) Geolocation signals only

They may later joined by a third "hybrid" mode where we can
experiment with how to blend the two (if that seems like it will provide
better results), or perform a fallback from one to the other.

Note that nothing is actually making Geolocation-based suggestions yet.
The shell command line can be used to simulate geolocation suggestions
in the meantime.

This commit doesn't contain real settings logic: the settings storage and
logic will be added in a follow-up commit.

Bug: 149014708
Test: atest com.android.server.timezonedetector
Test: atest core/tests/coretests/src/android/app/timezonedetector/
Change-Id: I83f8f64d40255567c035a67741b18ecaf6f0561d

Former-commit-id: 8164e67726c6ae438c85a1c60cced8c1e3c15f40