Use vibrator HAL controller for vibrator on and waveform
The new controller always triggers the callback, even if the HAL does
not support it. This requires changes in the VibratorService.java code
to rely on the native callback and use the ASYNC_TIMEOUT_MULTIPLIER for
the fallback.

This also changes the Vibration class to ignore native callbacks on
waveform effects, which are controlled by a separate VibrateThread that
triggers multiple vibrator ON calls (possibly indefinitely).

Bug: b/153418251
Test: atest FrameworksServicesTests:VibratorServiceTest
Change-Id: I09c9a0fd3f839c8f9eda65d96584a3e0294c6c21

Former-commit-id: bc5ff34bad0dae8306304816f66977a125a7657c