Fix the save UI does not restore
The new behavior of preesing the back key will move the activity
task to the background, but the activity has not yet been finished.
This will break some behaviors base on the onFinish() method when
the back key is pressed.
Make sure to restore the save UI after processing the back key event.

Fixes: 162203720
Test: atest CtsAutoFillServiceTestCases
Change-Id: I028056cadaa467cf1f795b0b76c88d24e0ba1955

Former-commit-id: add468c01205fd3ef4e0bd67d2c5662be289c0b5