Add KeyguardSliceViewController.
This removes view injection from KeyguardSliceView.

It also creates a KeyguardStatusView Dagger Subcomponent should be
initilialized whenever a KeyguardStatusView is inflated. From that
subcomponent, KeyguardClockSwitchController and
KeyguardSliceViewController can both be created and initialized,
passing their related views into their constructor.

This fixes a bug introduced earlier where a
KeyguardClockSwitchController was not being created in one place
where it was needed.

Bug: 162525274
Test: atest SystemUITests
Change-Id: Ie571cb0ca2b793c363e3c8ab397828f0ee939763

Former-commit-id: cd42c08702ef6cc25b974521ef9376e3b0d48bf8