Pass current PIP bounds to #updateReentryBounds.
When PIP exits, we attempt to save the current bounds so we know what
bounds to be used on PIP re-entry. This uses
PipTaskOrganizer#getLastReportedBounds, but this is not accessed in a
synchornized banner; it is possible that by the time it is accessed, it
is changed by something else (for example, PIP Animation). At the time
of #onPipTransitionStarted, just save the current PIP bounds and pass it
down.

Bug: 160721096
Test: Multiple re-entry of PIP result in the same bounds
Change-Id: I27a1ad3a2257333c9afedf13520f44406d07cd68

Former-commit-id: 1ced33a41cff0bd9d09a469bb8c9b42a79d566f0