Move PermissionsState into PermissionManagerService.
PermissionsState objects are now keyed by app IDs (UID without User
ID) inside PermissionManagerService. This assumes app IDs are stable
while the system is running, and across uninstallation of a system app
update, which is indeed the case. The PermissionStates inside package
setting is kept as-is because we need to migrate data from old
platforms.

To limit the scope of this change, persistence of permissions is also
left untouched, and we are temporarily telling
PermissionManagerService to write PermissionsState back into package
settings before persistence. This will be fixed once we move the
permission persistence out of package manager as well.

Bug: 158736025
Test: presubmit
Change-Id: I703fdbb9af64c30731b5e2c519e0b5d6c835d640

Former-commit-id: 7e8c82cdfe769b0beaa25ec1621ae2cc3f8e61eb