Filter out notif updates based on active media session
When an app is casting there can be two active media sessions, one local
and one remote. In this situation, there should only be a media object
for the remote session in QS. To achieve this, filter out updates for
the local session.

Bug: 158604810
Bug: 158813341
Bug: 158652134
Test: manual - cast from iHeartRadio. Toggle play/pause from media
object and check that device shown in output switcher isn't rapidly
changing.
Test: manual - cast from LivePhish. Check that only a single media
object appears in QS.

Change-Id: Ia7b8d2f3df3e68f63049e3443315fc26811efec4

Former-commit-id: 0e098127dbff682f968ef224687b8851e7e40f30