Introduce INPUT_CONSUMER permission for InputConsumer
Add the permission protect to createInputConsumer/destoryInputConsumer
in WindowManagerService to prevent it is possible for an app to consume
all the touch events and deny them from even reaching System UI.

Also fix the wallpaper inputconsumer should only show when we found a
visible wallpaper window.

Also limit 'createInputConsumer' to the defined types.

Test: atest PermissionPolicyTest
Test: ceate/destory an inputconsumer and test input
Fix: 162324374
Change-Id: I7a4c2c4cba7ea670873ab2ac49b9880d9a65cd86

Former-commit-id: 0d0ad1b60b3d9ed08b1df0b3a903af332568cd04