VoiceInteractionManagerService: userId vs. uid
A variable that (correctly) refers to a userId was
misnamed uid. Note that the function does return
a userId, and that the DBManager to which this
variable is passed requests a userHandle (which
is a synonym for userId). So it appears to have
just been a simple nisnaming of the variable.

Bug: 163626967
Test: compiles
Change-Id: I78b42ee3c3b79bb9a4a23219e9243201d0c0d440

Former-commit-id: 80209cfcfc934b3911c34e49aa161be8f4bd93fb