Implement UserTracker
This implements a new UserTracker for SystemUI that is meant to
centralize information about current user and current profiles.

It will serve the same function as CurrentUserContextTracker (replaced),
CurrentUserTracker (to be removed), and ACTION_USER_SWITCHED broadcasts
(to be removed). It'll also serve to remove all other calls to
ActivityManager.getCurrentUser().

Test: SystemUITests
Test: manual
Bug: 163579262

Change-Id: I6a7251485d88ce9b9131b9aff3b4343c3e8c823b

Former-commit-id: 0a7e5b85ae81d55eec8f9d2cfe5f69bf3f0503fc