Make framework-res dep in stubs explicit
The android.jar stubs includes the resource from framework-res. This
dependency was previously added implicitly inside the build system, but
using a java_genrule we can add it directly, which makes it easier to
make changes to this rule.

This change has the side-effect of adding the resources to some stubs
that did not previously include the resources (e.g. module-libs), which
I think is desired.

Bug: 161214753
Test: build sdk, diff out/dist before and after (shows resources added
      to some stubs)
Change-Id: I1829789af207cc45c49c0c68369fc7c0c085bcb0
Merged-In: I1829789af207cc45c49c0c68369fc7c0c085bcb0

Former-commit-id: 47994f37433e5f3e6f4522403e645ff0826e77ba