Reduce auto-cork delay to 50ms
Bug: 160971853

Reduce the auto-cork delay from 2s to 50ms.  Testing shows that this
introduces a very few new invalidate calls for the package_info cache
key, and greatly improves the hit ratio for the associated caches.

This change includes a fix to the cache-clear counter.  Some clears
were missing.

Test:
 * phone reboot
 * atest CtsContactsProviderTestCases
 * atest FrameworksServicesTests:UserSystemPackageInstallerTest
 * atest FrameworksServicesTests:PackageManagerSettingsTests
Verified that hit ratios improved over the baseline.  Verified that the
number of extra invalidates did not increase by more than 5.

Change-Id: I4b70b1da8c3927c74077e6da0e60dc56c2e007f1

Former-commit-id: 3da8ca16f4f62489b593ae5849109c827f36ebff