Remove use of INSTALL_DISABLE_VERIFICATION for staged install
Seems like INSTALL_DISABLE_VERIFICATION is now also being used to allow
test builds to optionally skip verification (introduced in ag/10057782).

As such, it is not possible to remove the flag completely anymore. I
removed the instances of the flag that were intended for staged
install flow. Now that we completely bypass verification using explicit
api, this check is no longer needed.

Bug: 10057782
Test: atest PackageVerifierTest
Test: atest StagedInstallTest
Change-Id: I7bc4df5c3afda59f6f37ea5379ee47c9ed8b09bb

Former-commit-id: c39590bf9f41fa537a741ef5f6b9fb54ae0befe3