Fix screenlayout calculation with overridden screenW/Hdp
Split-screen adjust-for-ime overrides screenW/Hdp in order
to prevent the app-configuration from changing. However, the
logic in Task#computeConfigResourceOverrides was ignoring
the overrides and blindly using the display's non-decor
bounds. This means that on loong screens, the app will
flip into/out-of long layout when adjusting (non-adjusted
task has no screenSizeDp overrides so it goes through logic
which crops mTmpNonDecorBounds).

Fix this by honoring the screenW/Hdp overrides when
calculating screenLayout.

Bug: 163848060
Test: On long screen (eg 21:9 aspect), open ime in second
      split. Also added regression unit-test.
Change-Id: I0fcec4f035e466fafedc31be5925c0b04a6580f7

Former-commit-id: ea5f48562f33acaa22deb2863b5fe2d6a3b11e42