Correct displayId info in log when WMS#addWindow returned ADD_INVALID_DISPLAY
As in WMS#addWindow will check if the application is accessible to add
window, if not, will print displayId value that application didn't have
access right on it.

Since getDisplayContentOrCreate(displayId, token) will return
the display of window token when the token exists,
if the displayId value is not same as the token display, the error
message may confuse the reader.

Bug: 163333691
Test: manual
Change-Id: Ic887fde9ab2e1f2b57a0c9241809287e193dcf48

Former-commit-id: 310125a2f94506b7e68e91765e6a928d33219a64