Use orientation to determine size compat bounds
If a display is resized to different orientation, the mapping between
rotation and orientation is also inverted. So it may be inaccurate to
calculate compat bounds according to current rotation. While the given
orientation is better because it is relative to the width and height.
That makes sure either the requested orientation takes effect or just
inherit parent's configuration.

Also consider the following cases:
 - Display rotation changed but activity orientation is not changed.
   The display is forced rotated and the activity is fixed orientation.
 - Display orientation changed but rotation is not changed.
   Resize fixed rotation display between portrait and landscape.

Fixes: 166822610
Test: SizeCompatTests#testFixedScreenBoundsWhenDisplaySizeChanged
Change-Id: I329325c134f37e08ecdb6311bff5cd331e3bc7c3

Former-commit-id: e3424aefa51d3e8587a90240099a937996c105ef