HexDumpTest: Fix assertion.
Commit 864256be58e48c14d49a8ca412ce682a3d260baa [formerly e121ca0a629580422fb4ca6af998f19e5b22f033] added test coverage,
including a helper function assertThrows(). The implementation of that
method was buggy because it only failed the test if it was throwing the
wrong kind of RuntimeException, instead of also when the code under test
wasn't throwing an exception at all.

This CL fixes the helper function. The test still passes (none of the
code under test was relying on the incorrect assertion).

Test: atest FrameworksCoreTests:com.android.internal.util.HexDumpTest

Change-Id: I8fe3e70ada8b880a03e124cb48e204e7bb880e57

Former-commit-id: 6bf181df87c0d782e095ab6617cbc4240f0cf418