Replace testWidevine23FpsH264Fixed with a shorter test.
Select WidevineYouTubePerformanceTests#testL3Cenc480P30
so the test does not exceed one minute during presubmit
testing.

Requires adding @Presubmit annotation to testL3Cenc480P30.
This is because some existing tests that use @Presubmit
annotation take precedent over the rest of the tests in
TEST_MAPPING file. If we do not use @Presubmit annotation,
the new test will not be executed. Using @Presubmit annotation
has the advantage of not having to explicitly specify
testL3Cenc480P30, in the include-filter.

Test: atest --test-mapping [directory with TEST_MAPPING]
  e.g. from frameworks/av, run "atest --test-mapping drm

Bug: 168159824
Change-Id: Ibd9d6df61b8f637b4e4cb1810ae5b595dce717ad

Former-commit-id: 6f0a63014600b5681159a23b3d86ae5419c23e50