Use normal rotation if orientation is changed by resume-scheduled app
With traditional normal rotation, if an app calls
Activity#setRequestedOrientation to change orientation in its onCreate
or onResume, the system display info will be updated immediately.
So if the app gets display rotation right after the orientation request,
it can get the rotated result.

While if fixed rotation transform takes effect for the case, though the
rotated display adjustments will be sent to the app, it is too late
that the app may have accessed the global display info which is not
rotated, that may cause inconsistent UI layout if the app depends on
the rotation info.

Bug: 167564038
Test: DisplayContentTests#testNoFixedRotationOnResumedScheduledApp
Change-Id: I3253264cb1bb8886074e88012582daa3589a1896

Former-commit-id: 4d576bfa25b764b3d2dd44761f632e7962443b02