Reapply display settings to DisplayContent when display changes.
When the physical display devices that map to a logical display change
onDisplayChanged() is called. However, the current logic only reads the
display metric overrides supplied from settings on creation of the
DisplayContent. This updates the logic to ensure that when a display
changes the most up-to-date settings values are read which may be
different from the current values if the logical display is mapped to a
different physical display.

Bug: 168808369
Test: atest:DisplayContentTests#testDisplaySettingsReappliedWhenDisplayChanged
Change-Id: Ia3a0d199f81e0e0ebc531d1edee54638946702f2

Former-commit-id: ecdee86b751a34bf24a992224ffb3fea69f22e56