Remove deprecated VibrationAttributes.getAudioAttrs method
Removing the deprecated AudioAttributes field from VibrationAttributes
in favor of a usage mapping to be used with AppOpsManager.checkAudioOp.

The mapping ensures the heuristic applied in VibrationAttributes.Builder
is taken into account by the VibratorManager.

This also adds a hidden method to Vibrator to allow forwarding a
VibrationAttributes to input device vibrators instead of relying on the
original AudioAttributes.

Bug: 163119475
Test: atest FrameworksServicesTests:VibratorServiceTest
      atest FrameworksCoreTests:VibratorTest
Change-Id: I83dbe079865b453c4cf42a6fd7e71f822acc39ea

Former-commit-id: 8f2674b2dca6f3927aeee265cc8d15999288fa78