Move defer-process-config for cached process to client side
Previously the cached state highly depends on the path of oom-adj
update, that causes potential lock contention between WMS and AMS.

Because the original major problem might be the heavy IO operations
by accessing the resources when handling configuration change from
dozens of non-important processes. The client side config deferring
could be enough help for the case.

This is also the last removal of holding WM lock in the path of
oom-adj update.

Bug: 143432064
Bug: 159104503
Test: ActivityThreadTest# \
      testHandleProcessConfigurationChanged_DependOnProcessState
Change-Id: Iefc9e02aa374bc74f263cf0c10612f89b864530b

Former-commit-id: e839c2ca794d5042fda8d4b69a7f918617030c58