Add debug logging to try to catch NPE in the wild.
~0.5% (or less) of our dogfooders are getting NPE's in the
KeyguardSliceProvider. It is unclear why this is happening and
we do not currently have reproduction steps for it.

Add some try/catch logging to the KeyguardSliceProvider to try
to catch this phenomenon so that we can narrow down why it is
occuring. This change can be reverted once we know more.

Bug: 168778439
Test: manual
Change-Id: Id65ee5bece75ef8497f7464079e5258781079fa9

Former-commit-id: 8daf476f54f5c5498f88b935a4823e6c5874f1ec