DO NOT MERGE Update the stable display frame for flexible insets
When using layoutStatusBar or layoutNavigationBar, the stable display
frame is modified to take the system bars into account. However when we
use flexible inset mapping, the stable display frame is not modified and
causes issues in getLayoutHint() and potentially affects how dialogs are
laid out.

This change ensures that even system bars that use flexible inset
mapping modifies the stable display frame as well.

Bug: 168846852
Test: manual, atest DialogFrameTests WindowInsetsControllerTests InsetsStateTest InsetsStateControllerTest
Change-Id: Ie89298eb98cf8c6bc9ada5494c544a3150b3b2de

Former-commit-id: 658f85c5c42bf2819683e786b24e6edc98065599