Update INTERNET perm state cache to use uids instead of appIds.
It really shouldn't have mattered whether we use uids or appIds for
this cache but given the way NetworkPolicy iterates over all apps
on the device (all_apps * all_users), it is possible that we end
up checking the permission state of apps which are not installed
on a user which will always be DENIED and we end up caching this.
So, we could end up treating an app as not having INTERNET permission
on a user even though it has.

Also, update the cache stragety to always check with PackageManager
when the permission state is denied in the cache just to be safe, until
NetworkPolicy iteration of apps is fixed.

Bug: 168299219
Test: atest cts/hostsidetests/net/src/com/android/cts/net/HostsideRestrictBackgroundNetworkTests.java
Test: manual

Change-Id: I6f2a60695a519a972c96ec8e053d3be5dc732461
Merged-In: I6f2a60695a519a972c96ec8e053d3be5dc732461

Former-commit-id: 97c29c74766cd6bee5d36792e52ba769b77b1406