Reverse the requested orientation on DisplayAreaGroup
When the DisplayAreaGroup has a different orientation from the Display,
reverse the app requested orientation.

As a result, when the display rotates to the reversed orientation, the
DAG and the requested app will be in the requested orientation.

For example, if the display is 1200x900 (landscape), and this DAG is
600x900 (portrait).
When an app below this DAG is requesting landscape, it should actually
request the display to be portrait, so that the DAG and the app will be
in landscape.

Bug: 155431879
Test: manual: test with dual display area policy
Test: atest WmTests:DisplayAreaGroupTest
Change-Id: Ie6874413ca0f6fa999632fb26f27e9913ccaee0d

Former-commit-id: 2d560e6976844cf5c5b785e6b6bf33d270251846