WindowManager: Remove windowsAreScaleable
This code was added for an earlier implementation of the PIP animation.
In this earlier implementation PIP was animated by directly scaling the
WSA level surface. This means that if we were in SCALING_MODE_FREEZE the
requested v active geometry logic in SF could freeze our
matrix/position/crop updates. Now that we manipulate the PIP surface
via higher levels of the hierarchy this wont be a problem and we don't
need to change the scaling mode.

Bug: 161937501
Test: Existing tests pass
Change-Id: Iaf10ee4fae9a56a79c1432d9ffd4c8227b832b01

Former-commit-id: f09bdfc4f76b026bae3b41ee5dd49e399a5f6454