5/N Move QS Tile tracking logic into QSPanelControllerBase.
This is a big, sort of hairy CL, but is a big push to completing this
series. One notable thing you see happen here is a lot less calls to
methods liks "setHost" which late-bind various member variables.
Classes such as QSTileHost are available to be injected at constrction
time and this CL starts taking advantage of that, removing the need
for a lot of null checks.

Bug: 168904199
Test: atest SystemUITests && manual
Change-Id: Ib4768d10e26debdea8d43daee61d2501ed250e67

Former-commit-id: 3834b11c62b2b396b4516480408373379b7889c5