Modify the user default for "geotz enabled"
This commit modifies the default for geolocation-based time zone
detection from "whatever the user's location toggle is" to just always
be off. This seems the most likely launch state as the current plan of
record is to not opt-in upgrading users, and to explicitly set the
setting in the SetUp Wizard for new users.

Bug: 151304765
Test: Build / TH only
Change-Id: Ifa5914b4d99708b54907259c50bd455897437225

Former-commit-id: 93f62ab79a50d8f2f3d610e3d22a2ace835f8df6