Do not allow uninstalling while an app is pinned
If an app is pinned we want to avoid ways to unpin without entering a
set passcode. If the package of the base activity in the pinned activity
stack is uninstalled then the device exits pinning mode so we want to
restrict uninstalling this package.

Bug: 135604684
Test: Pin test app, test app tries to uninstall itself
      Pin test app, `adb uninstall`
      Pin test app, test app launches second test app, assert that
          second test app can be uninstalled but base test app can't

Change-Id: I32ee438e9dd9e245bed6e6a9f4efd0abbb70de1f
Merged-In: I32ee438e9dd9e245bed6e6a9f4efd0abbb70de1f

Former-commit-id: eeeab037dfd3bbb59236766624e039c5cd0eea82