Make app installed cert flags persistent
Store AppInstalledRootCaCert and AppInstalledPrivateKey flags
in the config store to make them persistent. Resolves the issue
of certs that remain on the device after a connection is removed.
The reasone was that these flags were not persistent which caused
the logic to treat the certs as installed by the user.

Bug: 168871784
Test: Manually verify xml file contents
Test: atest XmlUtilTest
Change-Id: I06e93fa8688d1b483b68983d10b53fbe5acae32b

Former-commit-id: 806fd81a8249cdcc390e3068b07e8492ffa86cfa