Send transaction when possible for WSA
In WSA, we are sometimes using the global transaction. This will make
things difficult when trying to handle synced transactions for blast
sync. We need to make sure all changes on WSA that can happen during a
sync are put on the sync transaction so they can all be applied by the
client at the same time.

Also cleaned up some unused functions in SurfaceControl.

Test: Split and PIP
Change-Id: Ic1abc3c009c38d00593f80a4e03f895c026364e2

Former-commit-id: 3c57702ea1b060d668d32bf3f75ab1a4c7318c3a