PiP: Update DisplayInfo and movement bounds if display is different.
Although there are multiple places to watch out for display change in
PiP, when a task enters PiP, it's possible there has been no display
change, but simply a display addition. That being the case, it's
possible that PiP enters in a secondary display, and so we need to
update all the relevant information of the secondary display.

Bug: 133179171
Test: Enter PiP on a secondary display, default bounds is correct
Change-Id: I3eef2ef5c6b6be99105a876f7e8d43fbba39cbb2

Former-commit-id: 17b79c09b5e5ffc7754cc148a0eabba821db4311