Expose NetworkRegistrationInfo.getNrState() as public API
ConnectivityService is going to become a mainline module which
will not able to access hidden APIs. DataConnectionStat needs
to get the NR 5G state of the mobile data network for checking
if the network type is 5G or not. API in NetworkRegistrationInfo
is suggested to replace the original one in ServiceState. Thus,
expose it as a public API to support it.

Bug: 171183530
Test: make update-api ; m
Test: atest CtsTelephonyTestCases
Change-Id: I965499792286de4f1447559be66ca4be80bd0e6f

Former-commit-id: d32aa193fdee6c2be5e5e4cb67e80f57299f1c04