Allow voice recognition service to blame other apps access.
This is a workaround solution for Android R QPR2. We are not allowed
to change the API for QPR release. This is a special case that allows
the current voice recognizer to note proxy ops if it is also the
voice interactor. In S, we will define a new permission that designed
one as a trusted blamer and tie it to a role.

Bug: 17095434
Test: manual
Test: TreeHugger presubmit
Change-Id: I506bbeb95e622b99693dbf5a135c7961b7fd81c1

Former-commit-id: 38f832d4f356c2d735d7de68eb24ebea9d389089