Add genrule for combined srcjar of modules
The sources of the stubs are shipped as part of the public SDK.
Currently the SDK depends on the output of the monolithic metalava run,
but it can be assembled by merging the srcjars of the individual runs.

There is a diff -- module stub sources contain javadoc comments. I don't
think has any negative impact. Also package-info.html files are
included, which probably is fine.

Bug: 169034951
Test: diff new output with existing monolith
Change-Id: Idfc94cdb3317d3bde0b29fd8da0a9ccef47377f9

Former-commit-id: 5d0687a2efc759841aa1749ba5d69eeb4069834a